feat: update geofence modal and work location forms with improved titles and error handling
