+++ frontend/src/community/configurations/utils/geofenceUtils.ts
@@ -1,0 +1,4 @@
+export const MIN_RADIUS = 0;
+export const MAX_RADIUS = 300;
+
+export const formatRadius = (meters: number): string => `${meters}m`;
